Fixed nav and css and added all current models
Fixed nav links and combined css files so that one file can be used going forward. Additionally I added all the models we currently have available, and the headshots and designs content and copy. I added copy for the projects and team section on the home page. I commented where imgs are needed on home page. We are waiting on two imgs from photographer.